improve carousel
