+++ src/components/HeroCard.vue
@@ -66,6 +66,10 @@ function processCardType(type: Card['type']) {
     max-width: 480px;
     min-width: 275px;
 
+    @media (max-width: 400px) {
+        min-width: 0;
+    }
+
     border-radius: 1em;
     border: 1px solid #000;
