+++ src/stores/companion.ts
@@ -4,12 +4,13 @@ import { ref } from 'vue';
 
 export const useCompanionStore = defineStore('companion', () => {
     const id = useStorage<string | null>('goa2-faq.companion.id', null);
+    const focus = ref(2);
+
     const gold = useStorage<string | null>('goa2-faq.companion.gold', null);
     const silver = useStorage<string | null>('goa2-faq.companion.silver', null);
     const red = useStorage<string | null>('goa2-faq.companion.red', null);
     const green = useStorage<string | null>('goa2-faq.companion.green', null);
     const blue = useStorage<string | null>('goa2-faq.companion.blue', null);
-    const focus = ref(0);
 
     function resetCards() {
         gold.value = null;
@@ -21,8 +22,8 @@ export const useCompanionStore = defineStore('companion', () => {
 
     function reset() {
         id.value = null;
-        focus.value = 0;
         resetCards();
+        focus.value = 2;
     }
 
     return { id, gold, silver, red, green, blue, focus, reset, resetCards };
